Fixes various bugs in the Calc regarding Follow-ups
Thanks to YukariTea for the updated seals!

I've rewritten some part of the code. Rewriting it all could be really dangerous (bugs) and tiresome, so I decided to approach it little by little.

This "Little" update focuses on the follow-up logic and tries to fix it while making it smaller.